frameworks/google_translate/requirements.txt
boto3
google-cloud-translate
paramiko
requests
scp
